fix(meteo): preserve exact wind exceedance semantics and null legacy timestamps
Verifying the ingest-time reduction against the previous implementation
turned up two discrepancies.

reduceWindStep compared the unrounded float64 wind speed against the
25kt/34kt thresholds, whereas the original compared the float32 value it
stores. For speeds landing exactly on a threshold this flipped the
exceedance count, so a probability could disagree with the p50 sitting
next to it. Restores the original semantics.

CycleSummary tagged its timestamps omitempty, which encoding/json cannot
honour for a time.Time, so stores written before ingestion timing was
recorded reported "0001-01-01T00:00:00Z" through /debug/status. They are
pointers now and serialize as null.

With both fixed, the writer is byte-identical to the previous
implementation: generating deterministic, statistics-only and
full-ensemble stores under both versions produces the same 816 chunk
files and .zarray metadata, differing only in the wall-clock ingest
timestamps. The reader is unchanged, and a store stripped of the
timestamp fields still opens and queries correctly.

+++ services/meteo/internal/zarr/stats.go
@@ -160,11 +160,13 @@ func (sw *StoreWriter) reduceWindStep(step int) error {
 				if math.IsNaN(float64(u)) || math.IsNaN(float64(v)) {
 					continue
 				}
-				spd := math.Hypot(float64(u), float64(v))
-				sample[valid] = float32(spd)
+				// Threshold tests use the float32 speed that gets stored, not the wider
+				// intermediate, so a probability never disagrees with the p50 beside it.
+				spd := float32(math.Hypot(float64(u), float64(v)))
+				sample[valid] = spd
 				valid++
 				for i, t := range windThresholds {
-					if spd >= t.speed {
+					if float64(spd) >= t.speed {
 						exceed[i]++
 					}
 				}
